+++ dev/src/ts/scenes/game.ts
@@ -1,26 +1,26 @@
 export default class GameState extends Lightning.Scene {
 
-     protected particleEmitter:Lightning.ParticleEmitter
-     private _sprite:Lightnig.Sprite;
+    protected particleEmitter:Lightning.ParticleEmitter
+    private _sprite:Lightnig.Sprite;
 
-     private _start:boolean = false;
+    private _start:boolean = false;
 
     create() {
 
-        setTimeout(() => {
-            this.game.physics.lite.reset();
-            this.game.scene.start('menu');
-        }, 5000);
-
+        // setTimeout(() => {
+        //     this.game.physics.lite.reset();
+        //     this.game.scene.start('menu');
+        // }, 5000);
+        this.game.backgroundColour = 0xe5e5e5;
         this.game.physics.lite.enablePhysics();
-        let texture:Lightning.Texture =  Lightning.Geometry.Rect(5, 5).generateCanvasTexture();
-        let pool = this.game.physics.lite.createPool('test');
+        let texture:Lightning.Texture =  Lightning.Geometry.Rect(50, 50, 0xff22aa, 0.7, true);
+        let pool = this.game.physics.lite.createPool('test', true);
 
-        for(var i = 0; i < 30; i++) {
+        for(var i = 0; i < 7; i++) {
             let sprite = new Lightning.Sprite();
             sprite.texture = texture;
-            sprite.x = Lightning.Maths.rngFloat(0, this.game.width);            
-            sprite.y = Lightning.Maths.rngFloat(0, this.game.height);
+            sprite.x = Lightning.Maths.rngFloat(75, this.game.width -75);            
+            sprite.y = Lightning.Maths.rngFloat(75, this.game.height -75);
             this.add(sprite);
 
             sprite.enablePhysicsBody();
@@ -30,23 +30,23 @@ export default class GameState extends Lightning.Scene {
             sprite.body.velocity.x = Lightning.Maths.rngFloat(-5, 5);
             sprite.body.velocity.y = Lightning.Maths.rngFloat(-5, 5);
             sprite.body.collideOnWorldBounds = true;
-            sprite.body.gravityEnabled = true;
+            // sprite.body.gravityEnabled = true;
             sprite.body.restitution = 0.5;
         }
 
-        let bigGuy = new Lightning.Sprite();
-        bigGuy.texture = Lightning.Geometry.Rect(50, 50).generateCanvasTexture();
-        bigGuy.x = this.game.width / 2;
-        bigGuy.y = this.game.height / 2;
-        this.add(bigGuy);
+        // let bigGuy = new Lightning.Sprite();
+        // bigGuy.texture = Lightning.Geometry.Rect(50, 50, 0xff22aa, 1, true);
+        // bigGuy.x = this.game.width / 2;
+        // bigGuy.y = this.game.height / 2;
+        // this.add(bigGuy);
 
-        bigGuy.enablePhysicsBody();
-        bigGuy.body.enableDebug();
-        bigGuy.body.static = true;
+        // bigGuy.enablePhysicsBody();
+        // bigGuy.body.enableDebug();
+        // bigGuy.body.static = true;
 
-        let cEvent = this.game.physics.lite.createCollisionEvent('t', bigGuy.body, pool.bodies);
+        // let cEvent = this.game.physics.lite.createCollisionEvent('t', bigGuy.body, pool.bodies);
 
-        cEvent.onCollide(this.onCollide, this);
+        // cEvent.onCollide(this.onCollide, this);
 
         
         /**