+++ src/display/spritePool.ts
@@ -1,0 +1,22 @@
+/// <reference path="./../reference.d.ts" />
+
+namespace Lightning {
+    export class SpritePool extends PIXI.Container {
+
+        constructor() {
+            super();
+        }
+        
+        public spawn(setKey:string = null, count:number = 1, x?:number, y?:number) {
+
+        }
+
+        public setSpawnArea(rect:iRec) {
+            
+        }
+
+        public createCollection() {
+
+        }
+    }
+}