+++ src/physics/lite/manager.ts
@@ -12,6 +12,7 @@
  * Fix some bodies passing straight through
  *      I think this is because when a body becomes attached, the maximum amountn of events a being triggered
  * Apply mass to bodies
+ * move the maths stuff to the maths utils??
  * 
  */
 
@@ -30,7 +31,7 @@ namespace Lightning {
             super(game, true, true);
             this.game = game;
             this._worldBounds = { x: 0, y: 0, width: this.game.width, height: this.game.height };
-            this._gravity = {x: 0, y: 0.1};
+            this._gravity = {x: 0, y: 0};
         }
 
         /**
@@ -73,6 +74,7 @@ namespace Lightning {
         private mainUpdate() {
             for(let i in this._pools) {
                 for(let body of this._pools[i].bodies) {
+
                     if(body.collideOnWorldBounds) {
                         this.checkWorldCollide(body);
                     }
@@ -83,23 +85,25 @@ namespace Lightning {
                     
                     this.updatePosition(body);
                 }
+
+                this.checkPoolCollisions(this._pools[i]);
             }
 
-            for(let i in this._collisionEvents) {
-                this.checkCollisions(this._collisionEvents[i]);
+            // for(let i in this._collisionEvents) {
+            //     this.checkCollisions(this._collisionEvents[i]);
                 
-                for(let body of this._collisionEvents[i].bodies) {
-                    if(body.collideOnWorldBounds) {
-                        this.checkWorldCollide(body);
-                    }
+            //     for(let body of this._collisionEvents[i].bodies) {
+            //         if(body.collideOnWorldBounds) {
+            //             this.checkWorldCollide(body);
+            //         }
                     
-                    if(body.gravityEnabled) {
-                        this.calculateGravity(body);
-                    }
+            //         if(body.gravityEnabled) {
+            //             this.calculateGravity(body);
+            //         }
                     
-                    this.updatePosition(body);
-                }
-            }
+            //         this.updatePosition(body);
+            //     }
+            // }
 
             this.postUpdate();
         }
@@ -197,43 +201,84 @@ namespace Lightning {
         }
 
         private checkWorldCollide(body:LitePhysicsBody) {
-            //left
-            if(body.x <= this._worldBounds.x) {
-                body.velocity.x *= -1;
+            let x: number = body.x;
+            let y: number = body.y;
+
+            if(x <= this._worldBounds.x) {
+                body._isTouchingLeft = true;
+                if(body.velocity.x >= -0.1) {
+                    body.velocity.x = 0;
+                    body._isAsleepX = true;
+                } else {
+                    body.velocity.x *= -1;
+                }
+            } else {
+                body._isTouchingLeft = false;
             } 
-            //right
-            if(body.x >= this._worldBounds.width) {
-                body.velocity.x *= -1;
+            
+            if(x >= this._worldBounds.width) {
+                body._isTouchingRight = true;
+                if(body.velocity.x < 0.1) {
+                    body.velocity.x = 0;
+                    body._isAsleepX = true;
+                } else {
+                    body.velocity.x *= -1;
+                }
+            } else {
+                body._isTouchingRight = false;
+                body._isAsleepX = false;
+            } 
+            
+            if(y <= this._worldBounds.y) {
+                body._isTouchingUp = true;
+                if(body.velocity.y >= -0.1) {
+                    body.velocity.y = 0;
+                    body._isAsleepY = true;
+                } else {
+                    body.velocity.y *= -1;
+                }
+            } else {
+                body._isTouchingUp = false;
+                body._isAsleepY = false;
             }
-            //down
-            if(body.y <= this._worldBounds.height) {
-                body.velocity.y *= -1;
-            }
-            //up
-            if(body.y >= this._worldBounds.y) {
-                body.velocity.y *= -1;
+
+            if(y >= this._worldBounds.height) {
+                body._isTouchingDown = true;
+                if(body.velocity.y <= 0.1) {
+                    body.velocity.y = 0;
+                    body._isAsleepY = true;
+                } else {
+                    body.velocity.y *= -1;
+                }
+            } else {
+                body._isTouchingDown = false;
+                body._isAsleepY = false;
             }
         }
 
         private calculateGravity(body:LitePhysicsBody) {
             // checks if world gravity has already been applied
             if(body.deltaG.x === 0 && body.deltaG.y === 0) {
-                body.deltaG.x = this._gravity.x;
-                body.deltaG.y = this._gravity.y;
+                if(!body._isAsleepX) {
+                    body.deltaG.x = this._gravity.x;
+                }
+                if(!body._isAsleepY) {
+                    body.deltaG.y = this._gravity.y;
+                }
             }
         }
 
         private checkCollisions(collisionEvent:LitePhysicsCollisionEvent) {
             for(let i of collisionEvent.b1) {
                 for(let t of collisionEvent.b2) {
-                    if(this.AABBvsAABB(i, t) === true) {
-                        // check to see if the body's collision detection is paused
-                        if(!i.pauseCollisionDetection || !t.pauseCollisionDetection) {
-                            //this.resolveAABB(i, t);
-                            collisionEvent.collisionDetected(i, t);
-                        }
-                        return true;
-                    }
+                    // if(this.AABBvsAABB(i, t) === true) {
+                    //     // check to see if the body's collision detection is paused
+                    //     if(!i.pauseCollisionDetection || !t.pauseCollisionDetection) {
+                    //         this.resolveAABB(i, t);
+                    //         collisionEvent.collisionDetected(i, t);
+                    //     }
+                    //     return true;
+                    // }
                 }
             }
         }
@@ -249,44 +294,117 @@ namespace Lightning {
                     let body2 = pool.bodies[i];
 
                     // check two colisions here
-                    if(this.AABBvsAABB(body, body2) === true) {
-                        //this.resolveAABB(body, body2);
+                    let aabbResult = this.AABBvsAABB(body, body2);
+                    if(aabbResult === true) {
+                        // console.log(aabbResult)
+                        // this.resolveAABB(body, body2);
                     }       
                 }
                 c++;
             }
         }
 
-        private AABBvsAABB(b1:LitePhysicsBody, b2:LitePhysicsBody) {
-            if (b1.x < b2.x + b2.bounds.width &&
-                b1.x + b1.bounds.width > b2.x &&
-                b1.y < b2.y + b2.bounds.height &&
-                b1.bounds.height + b1.y > b2.y) {
-                    return true;
-            } else {
+        private AABBvsAABB(b1:LitePhysicsBody, b2:LitePhysicsBody): boolean {
+            let up: boolean, right: boolean, down: boolean, left: boolean = false;
+            let collision: boolean = false;
+
+            let aMinX: number = b1.x;
+            let aMaxX: number = b1.x + b1.bounds.width;
+            let bMinX: number = b2.x;
+            let bMaxX: number = b2.x + b2.bounds.width;
+
+            let aMinY: number = b1.y;
+            let aMaxY: number = b1.y + b1.bounds.height;
+            let bMinY: number = b2.y;
+            let bMaxY: number = b2.y + b2.bounds.height;
+
+            if(aMinX > bMaxX) {
                 return false;
             }
+
+            if(aMaxX < bMinX) {
+                return false;
+            }
+
+            if(aMinY > bMaxY) {
+                return false;
+            }
+
+            if(aMaxY < bMinY) {
+                return false;
+            }
+
+            let depthX: number = 0;
+            let depthY: number = 0;
+
+            let collisionMatrix = [0, 0, 0, 0, 0, 0, 0, 0];
+
+            if(aMinX < bMinX) {
+                collisionMatrix[1] = 1;
+                collisionMatrix[7] = 1;
+            } else {
+                collisionMatrix[5] = 1;
+                collisionMatrix[3] = 1;
+            }
+
+            if(aMinY < bMinY) {
+                collisionMatrix[0] = 1;
+                collisionMatrix[6] = 1;
+            } else {
+                collisionMatrix[2] = 1;
+                collisionMatrix[4] = 1;
+            }
+
+            let aCenterX: number = aMinX + (b1.bounds.width / 2);
+            let aCenterY: number = aMinY + (b1.bounds.height / 2);
+            let bCenterX: number = bMinX + (b2.bounds.width / 2);
+            let bCenterY: number = bMinY + (b2.bounds.height / 2);
+
+            // let a: number = aCenterX - bCenterX;
+            // let b: number = aCenterY - bCenterY;
+            // console.log(a, b);
+
+            depthX = Math.abs(aCenterX - bCenterX);
+            depthY = Math.abs(aCenterY - bCenterY);
+
+            if(depthX > depthY) {
+                b1.velocity.x *= -1;
+                b2.velocity.x *= -1;
+            } else {
+                b1.velocity.y *= -1;
+                b2.velocity.y *= -1;
+            }
+
+            return true;
+
+            // let fakeResult = {collision: false, up: false, right: false, down: false, left: false};
         }
 
         private resolveAABB(b1:LitePhysicsBody, b2:LitePhysicsBody) {
 
-            let b1KEx:number = 0.5 * b1.mass * (b1.velocity.x * b1.velocity.x);
-            let b1KEy:number = 0.5 * b1.mass * (b1.velocity.y * b1.velocity.y);
-            let b2KEx:number = 0.5 * b2.mass * (b2.velocity.x * b2.velocity.x);
-            let b2KEy:number = 0.5 * b2.mass * (b2.velocity.y * b2.velocity.y);
+            b1.velocity.x *= -1;
+            b1.velocity.y *= -1;
 
-            let diffx:number = Math.abs(b1KEx - b2KEx);
-            let diffy:number = Math.abs(b1KEy - b2KEy);
+            b2.velocity.x *= -1;
+            b2.velocity.y *= -1;
 
-            if(!b1.static) {
-                b1.velocity.x -= Math.sqrt(diffx / b2.mass);
-                b1.velocity.y -= Math.sqrt(diffy / b2.mass);
-            }
+            // let b1KEx:number = 0.5 * b1.mass * (b1.velocity.x * b1.velocity.x);
+            // let b1KEy:number = 0.5 * b1.mass * (b1.velocity.y * b1.velocity.y);
+            // let b2KEx:number = 0.5 * b2.mass * (b2.velocity.x * b2.velocity.x);
+            // let b2KEy:number = 0.5 * b2.mass * (b2.velocity.y * b2.velocity.y);
 
-            if(!b2.static) {
-                b2.velocity.x -= Math.sqrt(diffx / b1.mass);
-                b2.velocity.y -= Math.sqrt(diffy / b1.mass);
-            }
+            // let diffx:number = Math.abs(b1KEx - b2KEx);
+            // let diffy:number = Math.abs(b1KEy - b2KEy);
+
+            // if(!b1.static) {
+            //     b1.velocity.x -= Math.sqrt(diffx / b2.mass);
+            //     b1.velocity.y -= Math.sqrt(diffy / b2.mass);
+            // }
+
+            // if(!b2.static) {
+            //     b2.velocity.x -= Math.sqrt(diffx / b1.mass);
+            //     b2.velocity.y -= Math.sqrt(diffy / b1.mass);
+            // }
         }
 
         private outOfBounds(body:LitePhysicsBody) {